Move DefaultProjectState locking into dedicated synchronizer
Delete unused ModelContainerSynchronizer as well
