+++ subprojects/core/src/main/java/org/gradle/api/internal/project/DefaultProjectState.java
@@ -25,7 +25,6 @@ import org.gradle.api.project.IsolatedProject;
 import org.gradle.internal.build.BuildState;
 import org.gradle.internal.model.CalculatedModelValue;
 import org.gradle.internal.model.ModelContainer;
-import org.gradle.internal.model.ModelContainerSynchronizer;
 import org.gradle.internal.model.StateTransitionControllerFactory;
 import org.gradle.internal.project.ImmutableProjectDescriptor;
 import org.gradle.internal.resources.ProjectLeaseRegistry;
@@ -36,11 +35,9 @@ import org.jspecify.annotations.Nullable;
 
 import java.io.Closeable;
 import java.io.File;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -52,15 +49,12 @@ class DefaultProjectState implements ProjectState, Closeable {
     private final ImmutableProjectDescriptor descriptor;
     private final ProjectIdentity identity;
     private final IProjectFactory projectFactory;
+    private final ProjectStateSynchronizer synchronizer;
     private final ProjectLifecycleController controller;
     private final WorkerLeaseService workerLeaseService;
     private final ProjectStateLookup projectStateLookup;
 
-    private final ResourceLock allProjectsLock;
-    private final ResourceLock projectLock;
     private final ResourceLock taskLock;
-
-    private final Set<Thread> canDoAnythingToThisProject = new CopyOnWriteArraySet<>();
 
     DefaultProjectState(
         BuildState owner,
@@ -75,11 +69,10 @@ class DefaultProjectState implements ProjectState, Closeable {
         this.descriptor = descriptor;
         this.identity = descriptor.getIdentity();
         this.projectFactory = projectFactory;
-        this.controller = new ProjectLifecycleController(getDisplayName(), stateTransitionControllerFactory, new ModelContainerSynchronizer(this), buildServices);
+        this.synchronizer = new ProjectStateSynchronizer(owner, identity, workerLeaseService);
+        this.controller = new ProjectLifecycleController(getDisplayName(), stateTransitionControllerFactory, synchronizer, buildServices);
         this.workerLeaseService = workerLeaseService;
         this.projectStateLookup = projectStateLookup;
-        this.allProjectsLock = workerLeaseService.getAllProjectsLock(owner.getIdentityPath());
-        this.projectLock = workerLeaseService.getProjectLock(owner.getIdentityPath(), identity.getBuildTreePath());
         this.taskLock = workerLeaseService.getTaskExecutionLock(owner.getIdentityPath(), identity.getBuildTreePath());
     }
 
@@ -228,7 +221,7 @@ class DefaultProjectState implements ProjectState, Closeable {
 
     @Override
     public ResourceLock getAccessLock() {
-        return projectLock;
+        return synchronizer.getProjectLock();
     }
 
     @Override
@@ -251,47 +244,17 @@ class DefaultProjectState implements ProjectState, Closeable {
 
     @Override
     public <S extends @Nullable Object> S runWithModelLock(Supplier<S> action) {
-        Thread currentThread = Thread.currentThread();
-        if (workerLeaseService.isAllowedUncontrolledAccessToAnyProject() || canDoAnythingToThisProject.contains(currentThread)) {
-            // Current thread is allowed to access anything at any time, so run the action
-            return action.get();
-        }
-
-        Collection<? extends ResourceLock> currentLocks = workerLeaseService.getCurrentProjectLocks();
-        if (currentLocks.contains(projectLock) || currentLocks.contains(allProjectsLock)) {
-            // if we already hold the project lock for this project
-            if (currentLocks.size() == 1) {
-                // the lock for this project is the only lock we hold, can run the action
-                return action.get();
-            } else {
-                throw new IllegalStateException("Current thread holds more than one project lock. It should hold only one project lock at any given time.");
-            }
-        } else {
-            return workerLeaseService.withReplacedLocks(currentLocks, projectLock, action::get);
-        }
+        return synchronizer.withLock(action::get);
     }
 
     @Override
     public <S> S forceAccessToMutableState(Function<? super ProjectInternal, ? extends S> factory) {
-        Thread currentThread = Thread.currentThread();
-        boolean added = canDoAnythingToThisProject.add(currentThread);
-        try {
-            return factory.apply(getMutableModel());
-        } finally {
-            if (added) {
-                canDoAnythingToThisProject.remove(currentThread);
-            }
-        }
+        return synchronizer.forceLock(() -> factory.apply(getMutableModel()));
     }
 
     @Override
     public boolean hasMutableState() {
-        Thread currentThread = Thread.currentThread();
-        if (canDoAnythingToThisProject.contains(currentThread) || workerLeaseService.isAllowedUncontrolledAccessToAnyProject()) {
-            return true;
-        }
-        Collection<? extends ResourceLock> locks = workerLeaseService.getCurrentProjectLocks();
-        return locks.contains(projectLock) || locks.contains(allProjectsLock);
+        return synchronizer.holdsLock();
     }
 
     @Override
